# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Goth is a robust authentication API that supports both OAuth providers and traditional credentials-based authentication. It is designed to provide secure and flexible user authentication for various applications.",
        "title": "Goth",
        "contact": {
            "name": "Ankur Kumar",
            "url": "https://github.com/sirjager"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/user/{identity}": {
            "patch": {
                "description": "Partially Update User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "Update User Params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Identity can either be email or id",
                        "name": "identity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserResponse",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "description": "Fetch multiple users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Fetch Users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number: Default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per Page: Default 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UsersResponse",
                        "schema": {
                            "$ref": "#/definitions/UsersResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "get": {
                "description": "Refreshes Access Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Token",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "If true, returns User in body",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RefreshTokenResponse",
                        "schema": {
                            "$ref": "#/definitions/RefreshTokenResponse"
                        }
                    }
                }
            }
        },
...
        "/api/auth/user": {
            "get": {
                "description": "Get Authenticated User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Fetch",
                "responses": {
                    "200": {
                        "description": "UserResponse",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Initiate Authenticated User Deletion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code if already have",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "Partially Update User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "Update User Params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserResponse",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
...
        "/api/auth/{provider}": {
            "get": {
                "description": "Authenticates a user with a specified oauth provider",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "OAuth Provider",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth provider name [google,github]",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User object",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sessionID": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
...
        "UpdateUserParams": {
            "type": "object",
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                },
                "pictureURL": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "pictureURL": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "UserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
        "UsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}